requirements.txt
matplotlib==3.6.0
numpy==1.23.3
opencv-python==4.5.5.64
pandas==1.5.0
Pillow==9.2.0
PyYAML==6.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
seaborn==0.12.0
ipykernel==6.16.1
easyocr==1.6.2